Speed up server port bind speed
To do that we dynamic import the whole state/context logic
to let the server initialize first

The server now bind to port in 500ms, the bigger slowdown
now is inside minecraft-protocol: it's the generation of
the rsa key

Fix #306
