+++ src/player/chat.js
@@ -47,7 +47,7 @@ export function write_chat_msg({ world }, { message, client: { uuid } }) {
 }
 
 export default {
-  /** @type {import('../index.js').Observer} */
+  /** @type {import('../context.js').Observer} */
   observe({ client, get_state, world }) {
     client.on('chat', packet => {
       const { message } = packet
