+++ src/player/position.js
@@ -1,7 +1,7 @@
 const types = ['packet/position', 'packet/position_look', 'packet/look']
 
 export default {
-  /** @type {import('../index.js').Reducer} */
+  /** @type {import('../context.js').Reducer} */
   reduce(state, { type, payload }) {
     if (types.includes(type)) {
       return {
