+++ src/player/view_distance.js
@@ -1,5 +1,5 @@
 export default {
-  /** @type {import('../index.js').Reducer} */
+  /** @type {import('../context.js').Reducer} */
   reduce(state, { type, payload }) {
     if (type === 'packet/settings') {
       const { viewDistance } = payload
